% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strncc.R
\name{RcppNumberConnectedComponents}
\alias{RcppNumberConnectedComponents}
\title{Calculate the number of connected components from a vector of strings}
\usage{
RcppNumberConnectedComponents(strvectors, mm)
}
\arguments{
\item{strvectors}{A vector of strings}

\item{mm}{Maximal hamming distance allowed to add an edge between two strings}
}
\value{
The number of connected components (UMI)
}
\description{
Tokes in a vector of strings, and a threshold for number of mismatches allowed (hamming distance)
}
\examples{
RcppNumberConnectedComponents(c("ATTG","ACTG","CCGG"), 1)
dataframe \%>\% dplyr::group_by(sample) \%>\% dplyr::summarise(n_umi=Rncc::RcppNumberConnectedComponents(kmer, 3))
dataframe \%>\% multidplyr::partition(sample, cluster = cluster24) \%>\% dplyr::summarise(n_umi=Rncc::RcppNumberConnectedComponents(kmer, 3)) \%>\% collect()
}
